ECSRA-65 refactor: use interface instead of service implementation
